[Feat]: ReentrantLock을 이용한 동시성 제어 적용
Close #123
