+++ src/test/java/io/hhplus/tdd/point/PointServiceConcurrencyTest.java
@@ -1,0 +1,10 @@
+package io.hhplus.tdd.point;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class PointServiceConcurrencyTest {
+
+
+}
